Kavach_Setu_Face_Sign/Kavach_Setu_Face_Sign/face/google-cloud-credentials/requirements.txt
 deepface 
 opencv-python 
 ultralytics 
 google-cloud-vision
 fuzzywuzzy
 python-Levenshtein
 Pillow
 streamlit
 PyMuPDF==1.24.10
 uuid
 fastapi
 uvicorn
 python-dotenv
